+++ resolwe_bio/processes/differential_expression/edgeR.py
@@ -1,0 +1,201 @@
+"""Run EdgeR analysis."""
+from plumbum import TEE
+
+from resolwe.process import (
+    Cmd,
+    DataField,
+    FileField,
+    IntegerField,
+    JsonField,
+    ListField,
+    Persistence,
+    Process,
+    SchedulingClass,
+    StringField,
+)
+
+
+class EdgeR(Process):
+    """Run EdgeR analysis.
+
+    Empirical Analysis of Digital Gene Expression Data in R (edgeR).
+    Differential expression analysis of RNA-seq expression profiles with
+    biological replication. Implements a range of statistical methodology
+    based on the negative binomial distributions, including empirical Bayes
+    estimation, exact tests, generalized linear models and quasi-likelihood
+    tests. As well as RNA-seq, it be applied to differential signal analysis
+    of other types of genomic data that produce counts, including ChIP-seq,
+    Bisulfite-seq, SAGE and CAGE. See
+    [here](https://www.bioconductor.org/packages/devel/bioc/vignettes/edgeR/inst/doc/edgeRUsersGuide.pdf)
+    for more information.
+    """
+
+    slug = "differentialexpression-edger"
+    name = "edgeR"
+    process_type = "data:differentialexpression:edger"
+    version = "1.3.0"
+    category = "Differential Expression"
+    scheduling_class = SchedulingClass.BATCH
+    persistence = Persistence.CACHED
+    requirements = {
+        "expression-engine": "jinja",
+        "executor": {"docker": {"image": "resolwebio/rnaseq:4.9.0"}},
+        "resources": {"cores": 1, "memory": 8192},
+    }
+    data_name = "Differential expression (case vs. control)"
+
+    class Input:
+        """Input fields to process EdgeR."""
+
+        case = ListField(
+            DataField("expression"),
+            label="Case",
+            description="Case samples (replicates)",
+        )
+        control = ListField(
+            DataField("expression"),
+            label="Control",
+            description="Control samples (replicates)",
+        )
+
+        count_filter = IntegerField(
+            label="Raw counts filtering threshold",
+            default=10,
+            description="Filter genes in the expression matrix input. "
+            "Remove genes where the number of counts in all samples is "
+            "below the threshold.",
+        )
+
+    class Output:
+        """Output fields of the process EdgeR."""
+
+        raw = FileField("Differential expression")
+        de_json = JsonField(label="Results table (JSON)")
+        de_file = FileField(label="Results table (file)")
+        source = StringField(label="Gene ID database")
+        species = StringField(label="Species")
+        build = StringField(label="Build")
+        feature_type = StringField(label="Feature type")
+
+    def run(self, inputs, outputs):
+        """Run the analysis."""
+
+        for t in inputs.case:
+            if t in inputs.control:
+                self.error(
+                    "Case and Control groups must contain unique "
+                    f"samples. Sample {t.sample_name} is in both Case "
+                    "and Control group."
+                )
+
+        if len(inputs.case) < 2 or len(inputs.control) < 2:
+            self.error(
+                "Error in calculating edgeR dispersion, please provide more samples"
+            )
+
+        try:
+            case_paths = [c.rc.path for c in inputs.case]
+            control_paths = [c.rc.path for c in inputs.control]
+        except AttributeError:
+            self.error("Read counts are required when using edgeR")
+
+        conditions = ["case"] * len(case_paths) + ["control"] * len(control_paths)
+
+        expressions = inputs.case + inputs.control
+
+        for exp in expressions:
+            if exp.source != expressions[0].source:
+                self.error(
+                    "Input samples are of different Gene ID databases: "
+                    f"{exp.source} and {expressions[0].source}."
+                )
+            if exp.species != expressions[0].species:
+                self.error(
+                    "Input samples are of different Species: "
+                    f"{exp.species} and {expressions[0].species}."
+                )
+            if exp.build != expressions[0].build:
+                self.error(
+                    "Input samples are of different Panel types: "
+                    f"{exp.build} and {expressions[0].build}."
+                )
+            if exp.feature_type != expressions[0].feature_type:
+                self.error(
+                    "Input samples are of different Feature type: "
+                    f"{exp.feature_type} and {expressions[0].feature_type}."
+                )
+
+        for case in inputs.case:
+            if case in inputs.control:
+                self.error(
+                    "Case and Control groups must contain unique "
+                    f"samples. Sample {case.sample_name} is in both Case "
+                    "and Control group."
+                )
+
+        self.progress(0.1)
+
+        sample_files = case_paths + control_paths
+        merge_args = [
+            sample_files,
+            "--experiments",
+            sample_files,
+            "--intersection",
+            "--out",
+            "counts.tab",
+        ]
+        return_code, _, _ = Cmd["expressionmerge.py"][merge_args] & TEE(retcode=None)
+        if return_code:
+            self.error("Error merging read counts.")
+
+        filter_args = [
+            "-counts",
+            "counts.tab",
+            "-filter",
+            inputs.count_filter,
+            "-out",
+            "counts_filtered.tab",
+        ]
+        return_code, _, _ = Cmd["diffexp_filtering.R"][filter_args] & TEE(retcode=None)
+        if return_code:
+            self.error("Error while filtering read counts.")
+
+        args = [
+            "counts_filtered.tab",
+            "--sampleConditions",
+            conditions,
+        ]
+        return_code, _, _ = Cmd["run_edger.R"][args] & TEE(retcode=None)
+        if return_code:
+            self.error("Error computing differential expression (edgeR).")
+
+        self.progress(0.95)
+
+        edger_output = "diffexp_edgeR.tab"
+        args = [
+            edger_output,
+            "de_data.json",
+            "de_file.tab.gz",
+            "--gene_id",
+            "gene_id",
+            "--fdr",
+            "FDR",
+            "--pvalue",
+            "PValue",
+            "--logfc",
+            "logFC",
+        ]
+
+        return_code, _, _ = Cmd["parse_diffexp.py"][args] & TEE(retcode=None)
+        if return_code:
+            self.error(f"Error while parsing DGE results.")
+
+        (Cmd["gzip"][edger_output])()
+
+        outputs.raw = f"{edger_output}.gz"
+        outputs.de_json = "de_data.json"
+        outputs.de_file = "de_file.tab.gz"
+        outputs.source = expressions[0].source
+        outputs.species = expressions[0].species
+        outputs.build = expressions[0].build
+        outputs.feature_type = expressions[0].feature_type