+++ docs/CHANGELOG.rst
@@ -25,6 +25,7 @@ Changed
 - **BACKWARD INCOMPATIBLE:** Rewrite ``differentialexpression-deseq2``
   to Python
 - Add format parameter to ``macs2-callpeak``
+- Rewrite ``differentialexpression-edger`` to Python
 
 
 ===================